LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/transformations/2d.transformation.transform.nonfinite.worker-expected.txt


PASS transform() with Infinity/NaN is ignored 
